Updated cumulativecountytohrr.R to include moving growthrates, rolling averages, and ranks of each
The code now includes moving growth rates, along with a value "growthvalue" that sets the number of days growth rates are calculated over. Additionally, there are now rolling averages for cases, deaths, case rates, death rates, case growth rates, and case death rates. The number of days included in these rolling means can be changed by modifying "rollvalue", the alignment can be changed by modifying "alignvalue" to "left", "right", or "center, e.g setting the value to "left" means that the rolling average for that row would be calculated based on the average of the next "rollvalue" days. Finally, we include ranks for the growth rates and rolling averages, as well as an edit that extraneous header rows in outhrr.file